backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
app-store-web-scraper==0.2.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
config==0.5.1
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
google-play-scraper==1.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==2.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lazy-model==0.2.0
MarkupSafe==3.0.2
motor==3.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
openai==1.76.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
